+++ app/src/main/java/dk/shape/games/notifications/demo/mock/MockSubjectNotificationsParentFragment.kt
@@ -37,7 +37,7 @@ data class MockData(
 @Parcelize
 object SubjectNotificationsAction : Parcelable
 
-class SubjectNotificationsFragment : Fragment() {
+class MockSubjectNotificationsParentFragment : Fragment() {
 
     private val mockData = MockData(
         sportId = "football:0000",
